New Firebase configuration.
With this new way of injecting Firebase into the React App I can instantiate
a single instance of the firebase connection and set it as a context at the
top of the app.

This means that the firebase instance is a singleton and can be injected
into a component using a higher order component.